Auto adjust parameters, interrupt safe
This version of model performs an auto adjust on the parameters of the model. Since the training time could be quite long, an interrupt-safe mechanism is designed so that the program can restart from where it left off once interrupted.

The source file is divided into multiple files for enhanced readability.